requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
joblib==1.3.2
pandas==2.1.1
scikit-learn==1.4.2
pytest==7.0.1